end 396

+++ section25/api/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 const clearImage = (filePath) => {
   filePath = path.join(path.dirname(path.dirname(__filename)), filePath);
@@ -49,6 +50,7 @@ exports.createPost = (req, res, next) => {
     throw error;
   }
   
+  let creator;
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path.replace('\\', '/');
@@ -56,24 +58,24 @@ exports.createPost = (req, res, next) => {
     title: title, 
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'max'
-    }
+    creator: req.userId
   })
   post.save()
     .then(result => {
-        // Create post in db
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.push(post);
+      creator = user;
+      return user.save()
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: { 
-          _id: new Date().toISOString(), 
-          title: title, 
-          imageUrl: 'images/Protoss.png',
-          content: content,
-          creator: {
-            name: "max",
-          },
-          createdAt: new Date()
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name
         }
       });
     })
@@ -94,6 +96,7 @@ exports.getPostId = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      console.log('post', post)
       return res
         .status(200)
         .json({
@@ -132,6 +135,11 @@ exports.updatePost = (req, res, next) => {
           error.statusCode = 404;
           throw error;
         }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl != post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -165,9 +173,21 @@ exports.deletePost = (req, res, next) => {
       error.statusCode = 404;
       throw error;
     }
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error("Not authorized.");
+      error.statusCode = 403;
+      throw error;
+    }
     
     clearImage(post.imageUrl);
     return Post.deleteOne({_id: post})
+  })
+  .then(result => {
+    return User.findById(req.userId)
+  })
+  .then(user => {
+    user.posts.pull(req.params.postId);
+    return user.save();
   })
   .then(result => {
     res
